+++ MapboxSearch/autofill/src/main/java/com/mapbox/search/autofill/AddressAutofillOptions.kt
@@ -1,8 +1,6 @@
 package com.mapbox.search.autofill
 
 import android.os.Parcelable
-import com.mapbox.search.Country
-import com.mapbox.search.Language
 import kotlinx.parcelize.Parcelize
 import java.util.Locale
 
@@ -60,6 +58,84 @@ public class AddressAutofillOptions @JvmOverloads public constructor(
      */
     override fun toString(): String {
         return "AddressAutofillOptions(countries=$countries, language=$language)"
+    }
+
+    /**
+     * Country code to use in options.
+     * @property code country code in ISO 3166 alpha 2.
+     */
+    @Parcelize
+    public class Country(public val code: String) : Parcelable {
+
+        @JvmSynthetic
+        internal fun toCoreSdkType() = com.mapbox.search.Country(code)
+
+        /**
+         * @suppress
+         */
+        override fun equals(other: Any?): Boolean {
+            if (this === other) return true
+            if (javaClass != other?.javaClass) return false
+
+            other as Country
+
+            if (code != other.code) return false
+
+            return true
+        }
+
+        /**
+         * @suppress
+         */
+        override fun hashCode(): Int {
+            return code.hashCode()
+        }
+
+        /**
+         * @suppress
+         */
+        override fun toString(): String {
+            return "Country(code='$code')"
+        }
+    }
+
+    /**
+     * Preferred language of the autofill response.
+     * @property code language code in ISO 639-1.
+     */
+    @Parcelize
+    public class Language(public val code: String) : Parcelable {
+
+        @JvmSynthetic
+        internal fun toCoreSdkType() = com.mapbox.search.Language(code)
+
+        /**
+         * @suppress
+         */
+        override fun equals(other: Any?): Boolean {
+            if (this === other) return true
+            if (javaClass != other?.javaClass) return false
+
+            other as Language
+
+            if (code != other.code) return false
+
+            return true
+        }
+
+        /**
+         * @suppress
+         */
+        override fun hashCode(): Int {
+            return code.hashCode()
+        }
+
+        /**
+         * @suppress
+         */
+        override fun toString(): String {
+            return "Language(code='$code')"
+        }
     }
 
     private companion object {