+++ MapboxSearch/autofill/src/test/java/com/mapbox/search/autofill/AddressAutofillOptionsTest.kt
@@ -1,6 +1,5 @@
 package com.mapbox.search.autofill
 
-import com.mapbox.search.Language
 import com.mapbox.search.common.tests.ToStringVerifier
 import com.mapbox.test.dsl.TestCase
 import nl.jqno.equalsverifier.EqualsVerifier
@@ -17,7 +16,7 @@ internal class AddressAutofillOptionsTest {
 
                 Then(
                     "AddressAutofillOptions language should be determined by the current locale",
-                    expectedValue = Language(Locale.getDefault().language),
+                    expectedValue = AddressAutofillOptions.Language(Locale.getDefault().language),
                     actualValue = options.language
                 )
             }
@@ -27,9 +26,23 @@ internal class AddressAutofillOptionsTest {
     @TestFactory
     fun `Check AddressAutofillOptions's equals() and hashCode()`() = TestCase {
         Given("${AddressAutofillOptions::class.java.simpleName} class") {
-            When("Call equals() and hashCode()") {
+            When("Call AddressAutofillOptions equals() and hashCode()") {
                 Then("equals() and hashCode() should be implemented correctly") {
                     EqualsVerifier.forClass(AddressAutofillOptions::class.java)
+                        .verify()
+                }
+            }
+
+            When("Call AddressAutofillOptions.Language equals() and hashCode()") {
+                Then("equals() and hashCode() should be implemented correctly") {
+                    EqualsVerifier.forClass(AddressAutofillOptions.Language::class.java)
+                        .verify()
+                }
+            }
+
+            When("Call AddressAutofillOptions.Country equals() and hashCode()") {
+                Then("equals() and hashCode() should be implemented correctly") {
+                    EqualsVerifier.forClass(AddressAutofillOptions.Country::class.java)
                         .verify()
                 }
             }
@@ -39,10 +52,28 @@ internal class AddressAutofillOptionsTest {
     @TestFactory
     fun `Check AddressAutofillOptions's toString()`() = TestCase {
         Given("${AddressAutofillOptions::class.java.simpleName} class") {
-            When("Call toString()") {
+            When("Call AddressAutofillOptions toString()") {
                 Then("toString() function should be implemented correctly") {
                     ToStringVerifier(
                         clazz = AddressAutofillOptions::class,
+                        includeAllProperties = false
+                    ).verify()
+                }
+            }
+
+            When("Call AddressAutofillOptions.Language toString()") {
+                Then("toString() function should be implemented correctly") {
+                    ToStringVerifier(
+                        clazz = AddressAutofillOptions.Language::class,
+                        includeAllProperties = false
+                    ).verify()
+                }
+            }
+
+            When("Call AddressAutofillOptions.Country toString()") {
+                Then("toString() function should be implemented correctly") {
+                    ToStringVerifier(
+                        clazz = AddressAutofillOptions.Country::class,
                         includeAllProperties = false
                     ).verify()
                 }