+++ MapboxSearch/sdk/src/main/java/com/mapbox/search/ApiType.kt
@@ -19,7 +19,12 @@ public enum class ApiType {
      * Single Box Search API.
      */
     @Reserved(Flags.SBS)
-    SBS
+    SBS,
+
+    /**
+     * Autofill API.
+     */
+    AUTOFILL,
 }
 
 @JvmSynthetic
@@ -27,5 +32,6 @@ internal fun ApiType.mapToCore(): CoreApiType {
     return when (this) {
         ApiType.GEOCODING -> CoreApiType.GEOCODING
         ApiType.SBS -> CoreApiType.SBS
+        ApiType.AUTOFILL -> CoreApiType.AUTOFILL
     }
 }